adds project

+++ ft_printf/type_handlers/handle_hex.c
@@ -6,7 +6,7 @@
 /*   By: jakira-p <jakira-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/28 00:52:27 by jakira-p          #+#    #+#             */
-/*   Updated: 2021/09/29 04:27:54 by jakira-p         ###   ########.fr       */
+/*   Updated: 2021/09/29 18:43:12 by jakira-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	handle_hex(va_list args, t_metadata *data, int is_upper)
 	unsigned long	nbr;
 	char			*result;
 
-	nbr = va_arg(args, unsigned long);
+	nbr = va_arg(args, unsigned int);
 	if (is_upper)
 		result = to_hex(nbr, 1);
 	else
